My system was getting so much FreePBX lag...
I used to use the old school digium one and I could run this on a 512MB RAM VM, until the upgrade to 13 or 14, whichever Sangoma started at.

Anyways something in the existing code is going wonky there. ALL I know is that PHP's fopen does not return TRUE, only FALSE as Boolean. I figured if you only wanted to move on the script if you know for certain that the file handle returned FALSE, not both null as false as in !$fd would do, because it only uses double-type checking here. Test it out. Hopefully it works for you and solves a big use. I originally discovered the issue reducing the frequency of the various crontab jobs until I stopped receiving voice mail email notifications lol. For some reason I was not able to open crontab -e -u asterisk anymore so I was naturally forced to check the php….